moved all single file js to folder based #19
