+++ src/routes/ForgotPassword/components/ForgotPasswordForm/index.js
@@ -1,0 +1,3 @@
+import ForgotPasswordForm from './ForgotPasswordForm'
+
+export default ForgotPasswordForm
