+++ src/routes/ForgotPassword/components/ForgotPasswordView/index.js
@@ -1,0 +1,3 @@
+import ForgotPasswordView from './ForgotPasswordView'
+
+export default ForgotPasswordView
