+++ src/routes/ForgotPassword/containers/ForgotPasswordContainer/index.js
@@ -1,0 +1,3 @@
+import ForgotPasswordContainer from './ForgotPasswordContainer'
+
+export default ForgotPasswordContainer
