+++ src/routes/Login/components/LoginView/index.js
@@ -1,0 +1,3 @@
+import LoginView from './LoginView'
+
+export default LoginView
