+++ src/routes/Login/containers/LoginContainer/index.js
@@ -1,0 +1,3 @@
+import LoginContainer from './LoginContainer'
+
+export default LoginContainer
