+++ src/routes/Register/components/RegisterForm/index.js
@@ -1,0 +1,3 @@
+import RegisterForm from './RegisterForm'
+
+export default RegisterForm
