+++ src/routes/Register/components/RegisterView/index.js
@@ -1,0 +1,3 @@
+import RegisterView from './RegisterView'
+
+export default RegisterView
