+++ src/routes/Register/containers/RegisterContainer/index.js
@@ -1,0 +1,3 @@
+import RegisterContainer from './RegisterContainer'
+
+export default RegisterContainer
