+++ src/routes/Welcome/components/LinkedClientPanel/index.js
@@ -1,0 +1,3 @@
+import LinkedClientPanel from './LinkedClientPanel'
+
+export default LinkedClientPanel
