+++ src/routes/Welcome/components/WelcomeView/index.js
@@ -1,0 +1,3 @@
+import WelcomeView from './WelcomeView'
+
+export default WelcomeView
