+++ src/routes/Welcome/containers/WelcomeViewContainer/index.js
@@ -1,0 +1,3 @@
+import WelcomeViewContainer from './WelcomeViewContainer'
+
+export default WelcomeViewContainer
